src/modules/products/useCases/createProduct/CreateProductUseCase.ts
import { Product } from '@modules/products/infra/typeorm/entities/Product';
import { IProductsRepository } from '@modules/products/repositories/IProductsRepository';
import { AppError } from '@shared/errors/AppError';
import { inject, injectable } from 'tsyringe';

interface IRequest {
  name: string;
  description: string;
  daily_rate: number;
  license_plate: string;
  fine_amount: number;
  brand: string;
  category_id: string;
}

@injectable()
class CreateProductUseCase {
  constructor(
    @inject('ProductsRepository')
    private productsRepository: IProductsRepository,
  ) {}

  async execute({
    name,
    description,
    daily_rate,
    license_plate,
    fine_amount,
    brand,
    category_id,
  }: IRequest): Promise<Product> {
    //Verifica se a placa já foi cadastrada
    const productAlreadyExists = await this.productsRepository.findByLicensePlate(
      license_plate,
    );

    if (productAlreadyExists) {
      throw new AppError('Product already exists!');
    }

    const product = await this.productsRepository.create({
      name,
      description,
      daily_rate,
      license_plate,
      fine_amount,
      brand,
      category_id,
    });

    return product;
  }
}
export { CreateProductUseCase };
